Fjernet Screenlock-klasse og flyttet innhold til klassen TokenEncryption
Metoder var metoder knyttet til kryptering av token dersom bruker har aktiv skjermlås, og dermed virket det mer hensiktsmessig å ha de sammen resten.
